requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.49
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
pathspec==1.1.1
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyGithub==2.9.1
PyJWT==2.12.1
PyNaCl==1.6.2
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
smmap==5.0.3
sniffio==1.3.1
tiktoken==0.12.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
